debug: better upload flow, directly to email

+++ packages/app/features/settings/AppInfoScreen.tsx
@@ -1,5 +1,6 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
 import { useDebugStore } from '@tloncorp/shared';
+import { getCurrentUserId } from '@tloncorp/shared/dist/api';
 import * as store from '@tloncorp/shared/dist/store';
 import {
   AppSetting,
@@ -16,9 +17,10 @@ import {
 import { preSig } from '@urbit/aura';
 import * as Application from 'expo-application';
 import * as Updates from 'expo-updates';
-import { useEffect, useMemo } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { useCallback } from 'react';
 import { Alert, Platform, Switch } from 'react-native';
+import { getEmailClients, openComposer } from 'react-native-email-link';
 import { ScrollView } from 'react-native-gesture-handler';
 
 import { NOTIFY_PROVIDER, NOTIFY_SERVICE } from '../../constants';
@@ -30,16 +32,51 @@ const BUILD_VERSION = `${Platform.OS === 'ios' ? 'iOS' : 'Android'} ${Applicatio
 
 type Props = NativeStackScreenProps<RootStackParamList, 'AppInfo'>;
 
+function makeDebugEmail(appInfo: any, platformInfo: any) {
+  return `
+----------------------------------------------
+Insert description of problem here.
+----------------------------------------------
+
+Tlon ID: ${getCurrentUserId()}
+
+Platform Information:
+${JSON.stringify(platformInfo)}
+
+App Information:
+${JSON.stringify(appInfo)}
+`;
+}
+
 export function AppInfoScreen(props: Props) {
   const { data: appInfo } = store.useAppInfo();
-  const { enabled, logs, logsUrl, uploadLogs } = useDebugStore();
+  const { enabled, logs, logId, uploadLogs } = useDebugStore();
   const easUpdateDisplay = useMemo(() => getEasUpdateDisplay(Updates), []);
+  const [hasClients, setHasClients] = useState(true);
+
+  useEffect(() => {
+    async function checkClients() {
+      try {
+        const clients = await getEmailClients();
+        setHasClients(clients.length > 0);
+      } catch (e) {
+        setHasClients(false);
+      }
+    }
+
+    checkClients();
+  }, []);
 
   const onPressPreviewFeatures = useCallback(() => {
     props.navigation.navigate('FeatureFlags');
   }, [props.navigation]);
 
-  useEffect(() => {
+  const toggleDebugFlag = useCallback((enabled: boolean) => {
+    toggleDebug(enabled);
+    if (!enabled) {
+      return;
+    }
+
     Alert.alert(
       'Debug mode enabled',
       'Debug mode is now enabled. You may experience some degraded performance, because logs will be captured as you use the app. To get the best capture, you should kill the app and open it again.',
@@ -49,7 +86,24 @@ export function AppInfoScreen(props: Props) {
         },
       ]
     );
-  }, [enabled]);
+  }, []);
+
+  const onUploadLogs = useCallback(async () => {
+    const id = uploadLogs();
+
+    const { platform, appInfo } = useDebugStore.getState();
+    const platformInfo = await platform?.getDebugInfo();
+
+    if (!hasClients) {
+      return;
+    }
+
+    openComposer({
+      to: 'support@tlon.io',
+      subject: `${getCurrentUserId()} uploaded logs ${id}`,
+      body: makeDebugEmail(appInfo, platformInfo),
+    });
+  }, [hasClients]);
 
   return (
     <View flex={1}>
@@ -103,21 +157,22 @@ export function AppInfoScreen(props: Props) {
             <Switch
               style={{ flexShrink: 0 }}
               value={enabled}
-              onValueChange={toggleDebug}
+              onValueChange={toggleDebugFlag}
             ></Switch>
           </XStack>
 
           {enabled && logs.length > 0 && (
             <Stack>
-              <Button
-                onPress={() => {
-                  uploadLogs();
-                }}
-              >
+              <Button onPress={onUploadLogs}>
                 <Text>Upload logs ({logs.length})</Text>
               </Button>
-              <Text>{logsUrl}</Text>
             </Stack>
+          )}
+          {enabled && logId && !hasClients && (
+            <YStack padding="$l">
+              <Text>Please email support@tlon.io with this log ID:</Text>
+              <Text>{logId}</Text>
+            </YStack>
           )}
 
           <Stack marginTop="$xl">
